separate all components
