+++ sqstate/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+CURRENT_PATH = os.path.dirname(os.path.realpath(__file__))
